requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2
PyPDF2==3.0.1